+++ e3nn/o3/_tensor_product/_tensor_product.py
@@ -1,11 +1,12 @@
 from typing import Optional, List, Union
 
 import torch
+import torch.fx
 
 import e3nn
 from e3nn import o3
-from e3nn.util.codegen import CodeGenMixin
 from e3nn.util.jit import compile_mode
+from e3nn.util.codegen import CodeGenMixin
 from e3nn.util import prod
 
 from ._instruction import Instruction
@@ -255,7 +256,7 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         del opt_defaults
 
         # Generate the actual tensor product code
-        code_out, code_right, wigners = codegen_tensor_product(
+        graph_out, graph_right, wigners = codegen_tensor_product(
             self.irreps_in1,
             self.in1_var,
             self.irreps_in2,
@@ -268,11 +269,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
             self._specialized_code,
             self._optimize_einsums
         )
-
         self._codegen_register({
-            '_compiled_main_out': code_out,
-            '_compiled_main_right': code_right,
+            "_compiled_main_out": graph_out,
+            "_compiled_main_right": graph_right
         })
+
         self._wigners = wigners
 
         # === Determine weights ===